requirements.txt
streamlit
google-generativeai
wandb
